requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
PyYAML==6.0.1
schedule==1.2.0
